rabbitMQ/requirements.txt
pika-1.3.2